split node class into submodules
