+++ lib/dcell.rb
@@ -17,6 +17,9 @@ require 'dcell/messages'
 require 'dcell/sockets'
 require 'dcell/server'
 require 'dcell/node_manager'
+require 'dcell/node_communication'
+require 'dcell/node_rpc'
+require 'dcell/node_actors'
 require 'dcell/node'
 require 'dcell/global'
 require 'dcell/responses'
