+++ lib/dcell/node.rb
@@ -10,12 +10,19 @@ module DCell
     include Celluloid
     include Celluloid::FSM
 
+    include DCell::Node::Communication
+    include DCell::Node::RPC
+    include DCell::Node::Actors
+
     attr_reader :id
+
     finalizer :shutdown
 
     # FSM
     default_state :disconnected
-    state :shutdown
+    state :shutdown do
+      on_shutdown
+    end
     state :disconnected, to: [:connected, :shutdown]
     state :connected do
       on_connected
@@ -32,18 +39,12 @@ module DCell
     end
 
     def initialize(id, addr, server=false)
-      @id, @addr = id, addr
-      @requests = ResourceManager.new
-      @actors = ResourceManager.new
-      @remote_dead = false
-      @leech = false
-      @socket, @rsocket, @raddr = nil, nil, nil
-      @ttl, @rserver = nil, nil
+      @id = id
 
-      @heartbeat_rate    = DCell.heartbeat_rate
-      @heartbeat_timeout = DCell.heartbeat_timeout
-      @request_timeout   = DCell.request_timeout
-      @ttl_rate          = DCell.ttl_rate
+      init_rpc
+      init_actors
+      init_defaults
+      init_comm addr
 
       if server
         update_ttl
@@ -89,187 +90,41 @@ module DCell
     # Internal API
     ##################################################
 
-    def save_request(request)
-      @requests.register(request.id) { request }
-    end
+    def init_defaults
+      @ttl = nil
 
-    def delete_request(request)
-      @requests.delete request.id
-    end
-
-    def cancel_requests
-      @requests.each do |id, request|
-        address = request.sender.address
-        rsp = CancelResponse.new id, address
-        rsp.dispatch
-      end
-    end
-
-    def add_actor(actor)
-      @actors.register(actor.object_id) { actor }
-    end
-
-    def kill_actors
-      @actors.clear do |id, actor|
-        begin
-          actor.terminate
-        rescue Celluloid::DeadActorError
-        end
-      end
+      @heartbeat_rate    = DCell.heartbeat_rate
+      @heartbeat_timeout = DCell.heartbeat_timeout
+      @request_timeout   = DCell.request_timeout
+      @ttl_rate          = DCell.ttl_rate
     end
 
     def detach
       kill_actors
       cancel_requests
-      @remote_dead = true
+      remote_dead
       terminate
-    end
-
-    def close_sockets
-      [@socket, @rsocket, @rserver].each do |socket|
-        next unless socket && socket.alive?
-        socket.terminate
-      end
-    end
-
-    def self_cleanup
-      NodeCache.delete @id
-      MailboxManager.delete Thread.mailbox
-      instance_variables.each { |iv| remove_instance_variable iv }
-    end
-
-    def say_goodbye
-      return if @remote_dead || DCell.id == @id
-      kill_actors
-      farewell
     end
 
     # Graceful termination of the node
     def shutdown
       transition :shutdown
-      Logger.info "Disconnected from #{@id}"
-      say_goodbye
-      close_sockets
-      self_cleanup
-    end
-
-    # Obtain socket for relay messages
-    def rsocket
-      return @rsocket if @rsocket
-      send_relayopen unless @raddr
-      @rsocket = ClientServer.new @raddr, @heartbeat_timeout * 1000
-    end
-
-    # Obtain socket for management messages
-    def socket
-      return @socket if @socket
-      @socket = ClientServer.new @addr, @heartbeat_timeout * 1000
-      @socket.farewell = true
-      transition :connected
-      @socket
-    end
-
-    # Pack and send a message to another DCell node
-    def send_message(message, pipe=:request)
-      queue = nil
-      if pipe == :request
-        queue = socket
-      elsif pipe == :response
-        queue = Celluloid::Actor[:server]
-      elsif pipe == :relay
-        queue = rsocket
-      end
-
-      begin
-        message = message.to_msgpack
-      rescue => e
-        abort e
-      end
-      queue.write @id, message
-    end
-
-    # Send request and wait for response
-    def push_request(request, pipe=:request, timeout=@request_timeout)
-      send_message request, pipe
-      save_request request
-      response = receive(timeout) do |msg|
-        msg.respond_to?(:request_id) && msg.request_id == request.id
-      end
-      delete_request request
-      abort NoResponseError.new unless response
-      response
-    end
-
-    # Send request and handle unroll response
-    def send_request(request, pipe=:request, timeout=nil)
-      response = push_request request, pipe, timeout
-      return if response.is_a? CancelResponse
-      if response.is_a? ErrorResponse
-        value = response.value
-        klass = Utils.full_const_get value[:class]
-        exception = klass.new value[:msg]
-        exception.set_backtrace value[:tb]
-        abort exception
-      end
-      response.value
-    end
-
-    # Relay message to remote actor
-    def relay(message)
-      request = Message::Relay.new(Thread.mailbox, message)
-      send_request request, :relay
-    end
-
-    # Relay async message to remote actor
-    def async_relay(message)
-      request = Message::Relay.new(Thread.mailbox, message)
-      send_message request, :relay
-    end
-
-    # Goodbye message to remote actor
-    def farewell
-      return unless Directory[@id].alive?
-      request = Message::Farewell.new
-      send_message request
-    rescue
-    end
-
-    # Send a heartbeat message after the given interval
-    def send_heartbeat
-      return if DCell.id == @id
-      request = DCell::Message::Heartbeat.new @id
-      send_message request, @leech ? :response : :request
-      @heartbeat = after(@heartbeat_rate) { send_heartbeat }
-    end
-
-    # Handle an incoming heartbeat for this node
-    def handle_heartbeat(from)
-      return if from == @id
-      @leech = true unless state == :connected
-      transition :connected
-      transition :partitioned, delay: @heartbeat_timeout
-    end
-
-    # Send an advertising message
-    def send_relayopen
-      request = Message::RelayOpen.new(Thread.mailbox)
-      @raddr = send_request request
-    end
-
-    # Handle an incoming node advertising message for this node
-    def handle_relayopen
-      @rsocket = rserver
-    end
-
-    def rserver
-      return @rserver if @rserver
-      @rserver = RelayServer.new
+      farewell
+      kill_actors
+      close_comm
+      NodeCache.delete @id
+      MailboxManager.delete Thread.mailbox
+      instance_variables.each { |iv| remove_instance_variable iv }
     end
 
     # Update TTL in registry
     def update_ttl
       Directory[@id].update_ttl
       @ttl = after(@ttl_rate) { update_ttl }
+    end
+
+    def on_shutdown
+      Logger.info "Disconnected from #{@id}"
     end
 
     def on_connected
